+++ android-gif-drawable/src/main/java/pl/droidsonroids/gif/GifDrawable.java
@@ -648,19 +648,30 @@ public class GifDrawable extends Drawable implements Animatable, MediaPlayerCont
 	}
 
 	/**
-	 * Returns size of the allocated memory used to store pixels of this object.
-	 * It counts length of all frame buffers. Returned value does not change during runtime.
+	 * Returns size of the memory needed to store pixels of this object. It counts possible length of all frame buffers.
+	 * Returned value may be lower than amount of actually allocated memory if GIF uses dispose to previous method but frame requiring it
+	 * has never been needed yet. Returned value does not change during runtime.
 	 *
-	 * @return size of the allocated memory used to store pixels of this object
+	 * @return possible size of the memory needed to store pixels of this object
 	 */
 	public long getAllocationByteCount() {
 		long byteCount = mNativeInfoHandle.getAllocationByteCount();
 		if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
 			byteCount += mBuffer.getAllocationByteCount();
 		} else {
-			byteCount += mBuffer.getRowBytes() * mBuffer.getHeight();
+			byteCount += getFrameByteCount();
 		}
 		return byteCount;
+	}
+
+	/**
+	 * Returns the maximum possible size of the allocated memory used to store pixels and metadata of this object.
+	 * It counts length of all frame buffers. Returned value does not change over time.
+	 *
+	 * @return maximum possible size of the allocated memory used to store pixels and metadata of this object
+	 */
+	public long getMetadataAllocationByteCount() {
+		return mNativeInfoHandle.getMetadataByteCount();
 	}
 
 	/**